pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "myyahoomails"
version = "0.1.0"
description = "Email analysis platform for Yahoo Mail via IMAP"
requires-python = ">=3.9"
dependencies = [
    # IMAP & email parsing
    "imapclient>=2.3",
    "mail-parser>=3.15",
    "chardet>=5.0",
    # Database
    # (sqlite3 is stdlib — no extra dep needed)
    # CLI
    "click>=8.1",
    "rich>=13.0",
    "tqdm>=4.66",
    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # LLM providers
    "anthropic>=0.40",
    "openai>=1.50",
    "groq>=0.11",
    "httpx>=0.27",
    # Reports
    "python-docx>=1.1",
    "weasyprint>=62.0",
    "matplotlib>=3.8",
    # Web dashboard (Phase 5)
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
mye = "cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
